+++ types/index.d.ts
@@ -18,6 +18,9 @@ interface WebPageIframe {
 interface MainPanel {
     scrollTo(block: Block|null, behavior: 'smooth'|'auto' = 'smooth');
     getEl(): HTMLElement;
+    registerSection(name: String, Cls: preact.AnyComponent): void;
+    getSection(name: String): preact.AnyComponent;
+    getSections(): Map<preact.AnyComponent>;
 }
 
 interface BlockTypes {